Update and rename app.js to app-beast2.js
My solution to Beast 2 for Watch & Code. 

This answer is a bit braindead, just replacing the call to util.store in render with like, six calls in all the other functions where the logical todos array is modified (so, toggle, toggleAll, destroy, destroyCompleted, update, and create). 

However, as best I can tell, it works, and is not hard to understand, so I think it's OK. 

Will update or modify if this solution proves unsatisfactory.